+++ CommonplaceBookApp/EmptyViewController.swift
@@ -1,0 +1,6 @@
+// Copyright © 2019 Brian's Brain. All rights reserved.
+
+import UIKit
+
+/// Empty detail view
+final class EmptyViewController: UIViewController { }